# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for get cryptocurrencies",
        "title": "My-Portfolio",
        "contact": {
            "name": "My-Portfolio support"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8000",
    "paths": {
...
        "/api/v1/login": {
            "post": {
                "description": "авторизация",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "parameters": [
                    {
                        "description": "data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_controllers.LoginData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_app_controllers.Tokens"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_app_controllers.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_app_controllers.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/v1/portfolio": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "портфель",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "portfolio"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_app_controllers.PortfolioGetResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_app_controllers.ResponseError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "добавить в портфель криптовалюту",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "portfolio"
                ],
                "parameters": [
                    {
                        "description": "data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_controllers.PortfolioPostData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_app_controllers.ResponseSuccess"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_app_controllers.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/v1/portfolio/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "удалить криптовалюту из портфеля",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "portfolio"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Portfolio detail ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_app_controllers.PortfolioGetResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_app_controllers.ResponseError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "обновитель информацию о криптовалюте в портфеле",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "portfolio"
                ],
                "parameters": [
                    {
                        "description": "data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_controllers.PortfolioPatchData"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Portfolio detail ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_app_controllers.ResponseSuccess"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_app_controllers.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/v1/refresh": {
            "post": {
                "description": "обновление токенов",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "parameters": [
                    {
                        "description": "data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_controllers.RefreshData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_app_controllers.Tokens"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_app_controllers.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_app_controllers.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/v1/register": {
            "post": {
                "description": "регистрация",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "parameters": [
                    {
                        "description": "data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_controllers.RegisterData"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_app_controllers.Tokens"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_app_controllers.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_app_controllers.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_app_controllers.LoginData": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "internal_app_controllers.PortfolioDetailGetResponse": {
            "type": "object",
            "required": [
                "count",
                "price",
                "purchase_time"
            ],
            "properties": {
                "ROI": {
                    "type": "number"
                },
                "commentary": {
                    "type": "string"
                },
                "count": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "cryptocurrency": {
                    "type": "string"
                },
                "cryptocurrency_id": {
                    "type": "integer"
                },
                "cryptocurrency_symbol": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "now_price": {
                    "type": "number"
                },
                "now_sum": {
                    "type": "number"
                },
                "percent_change_24h": {
                    "type": "number"
                },
                "percent_change_30d": {
                    "type": "number"
                },
                "percent_change_90d": {
                    "type": "number"
                },
                "portfolio_share": {
                    "type": "number"
                },
                "price": {
                    "type": "number"
                },
                "purchase_time": {
                    "type": "string"
                },
                "sum": {
                    "type": "number"
                }
            }
        },
        "internal_app_controllers.PortfolioGetResponse": {
            "type": "object",
            "properties": {
                "ROI": {
                    "type": "number"
                },
                "cryptocurrencies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_app_controllers.PortfolioDetailGetResponse"
                    }
                },
                "total_now_sum": {
                    "type": "number"
                },
                "total_purchased_sum": {
                    "type": "number"
                }
            }
        },
        "internal_app_controllers.PortfolioPatchData": {
            "type": "object",
            "properties": {
                "commentary": {
                    "type": "string"
                },
                "count": {
                    "type": "number"
                },
                "price": {
                    "type": "number"
                },
                "purchase_time": {
                    "type": "string"
                }
            }
        },
        "internal_app_controllers.PortfolioPostData": {
            "type": "object",
            "required": [
                "cmc_cryptocurrency_id",
                "count",
                "price",
                "purchase_time"
            ],
            "properties": {
                "cmc_cryptocurrency_id": {
                    "type": "integer"
                },
                "commentary": {
                    "type": "string"
                },
                "count": {
                    "type": "number"
                },
                "price": {
                    "type": "number"
                },
                "purchase_time": {
                    "type": "string"
                }
            }
        },
        "internal_app_controllers.RefreshData": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "internal_app_controllers.RegisterData": {
            "type": "object",
            "required": [
                "birth_date",
                "email",
                "first_name",
                "gender",
                "last_name",
                "password"
            ],
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string",
                    "minLength": 2
                },
                "gender": {
                    "type": "string",
                    "enum": [
                        "male",
                        "female"
                    ]
                },
                "last_name": {
                    "type": "string",
                    "minLength": 2
                },
                "password": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 8
                }
            }
        },
        "internal_app_controllers.ResponseError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "boolean"
                },
                "error_string": {
                    "type": "string"
                }
            }
        },
        "internal_app_controllers.ResponseSuccess": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "boolean"
                }
            }
        },
...
        "internal_app_controllers.Tokens": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}